Making Functionality Improvements
You can now get a random pokemon when clicking the pokeball. This was implemented using a function that gets a random pokemon number and passes it in to the existing pokemon search function. This involved creating a seperate function for the normal search that passes the value into the pokemon Search function.

Also made changes to the UI to make the Pokemon Card look better and implemented searching on the press of the enter key.